tooling/zelensky_telegram/requirements.txt
python-dotenv==0.20.0
Telethon==1.24.0
